refactor: CardInputLayout 폴더 위치 변경, description 추가
