+++ src/components/common/CardInputLayout/index.tsx
@@ -6,10 +6,20 @@ import { Flex } from '@/components/common/Flex';
 import { Text } from '@/components/common/Text';
 
 export type Props = {
+  /**
+   * Sets the title for the card input field.
+   */
   headerText: string;
+  /**
+   * Sets the description for the card input field.
+   */
   description?: string;
+  /**
+   * Sets the label for the card input field.
+   */
   label: string;
 } & PropsWithChildren;
+
 export const CardInputLayout = ({ headerText, description, label, children }: Props) => {
   return (
     <StyledCardInputLayoutContainer>
